Cleanup source code
