+++ extension-manager-integration-test-java/src/test/java/com/exasol/extensionmanager/itest/IntegrationTestCommon.java
@@ -1,0 +1,8 @@
+package com.exasol.extensionmanager.itest;
+
+import java.nio.file.Path;
+import java.nio.file.Paths;
+
+class IntegrationTestCommon {
+    static final Path TESTING_EXTENSION_SOURCE_DIR = Paths.get("testing-extension").toAbsolutePath();
+}
